restructure the tests directory
