+++ server/src/tests/testHandle/car.test.js
@@ -3,7 +3,7 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 
-import server from '../index';
+import server from '../../index';
 
 chai.should();
 
@@ -44,7 +44,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', 'used')
         .field('price', 1750000)
         .field('manufacturer', 'bmw')
@@ -90,7 +90,7 @@ describe('Test Suite For Car Endpoints', () => {
         .request(server)
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
-        .attach('image', 'server/src/index.js')
+        .attach('image', './server/src/index.js')
         .type('form')
         .field('state', 'used')
         .field('price', 1750000)
@@ -111,7 +111,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', '')
         .field('price', 1750000)
         .field('manufacturer', 'bmw')
@@ -130,7 +130,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', 122)
         .field('price', 1750000)
         .field('manufacturer', 'bmw')
@@ -149,7 +149,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', 'boy')
         .field('price', 1750000)
         .field('manufacturer', 'bmw')
@@ -168,7 +168,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', 'used')
         .field('price', '')
         .field('manufacturer', 'bmw')
@@ -187,7 +187,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', 'used')
         .field('price', -173663376)
         .field('manufacturer', 'bmw')
@@ -206,7 +206,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', 'used')
         .field('price', 'amount')
         .field('manufacturer', 'bmw')
@@ -225,7 +225,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', 'used')
         .field('price', 1750000)
         .field('manufacturer', '')
@@ -245,7 +245,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', 'used')
         .field('price', 1750000)
         .field('manufacturer', 'bmw')
@@ -265,7 +265,7 @@ describe('Test Suite For Car Endpoints', () => {
         .post('/api/v1/car')
         .set('Authorization', `Bearer ${buyerSellerToken}`)
         .type('form')
-        .attach('image', 'server/src/tests/fortest.png')
+        .attach('image', './server/src/tests/lib/fortest.png')
         .field('state', 'used')
         .field('price', 1750000)
         .field('manufacturer', 'bmw')
